Move apicalls to Game
