+++ src/components/Start/Start.js
@@ -7,9 +7,10 @@ const Start = (props) => {
   return (
     <section  className="rules-container">
       <h2 className="headline" >RULES</h2>
-      <p>You will be given a quote from a character in Breaking Bad and three characters</p>
-      <p>Choose the character that you think authored the quote</p> 
-      <h3>Be the one who knocks to start a game</h3>
+      <p>You will be given 10 quotes from Breaking Bad or Better Call Saul and 3 characters.</p>
+      <p>Choose the character that you think authored the quote.</p> 
+      <p>Refreshing or leaving the page will start a new game.</p>
+      <h3>Be the one who knocks to start a game.</h3>
       
       <button 
         onClick={() => startGame()}
